requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
ecdsa==0.14.1
fastapi==0.61.1
fastapi-utils==0.2.1
h11==0.9.0
httptools==0.1.1
idna==2.10
mysql-connector-python==8.0.21
passlib==1.7.2
protobuf==3.13.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyMySQL==0.10.1
pyserial==3.4
python-jose==3.2.0
requests==2.24.0
responses==0.12.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
